+++ paddle/fluid/distributed/ps/table/sparse_sgd_rule.cc
@@ -88,7 +88,7 @@ void SparseAdaGradSGDRule::UpdateValueWork(float* w, float* sgd,
   float& g2sum = sgd[G2SumIndex()];
   double add_g2sum = 0;
 
-  for (int i = 0; i < _embedding_dim; i++) {
+  for (size_t i = 0; i < _embedding_dim; i++) {
     double scaled_grad = grad[i] / scale;
     w[i] -= learning_rate_ * scaled_grad *
             sqrt(_initial_g2sum / (_initial_g2sum + g2sum));
@@ -101,7 +101,7 @@ void SparseAdaGradSGDRule::UpdateValueWork(float* w, float* sgd,
 
 void SparseAdaGradSGDRule::InitValueWork(float* value, float* sgd,
                                          bool zero_init) {
-  for (int i = 0; i < _embedding_dim; ++i) {
+  for (size_t i = 0; i < _embedding_dim; ++i) {
     if (zero_init) {
       value[i] = 0.0;
       BoundValue(value[i]);
@@ -139,7 +139,7 @@ void StdAdaGradSGDRule::LoadConfig(const SparseCommonSGDRuleParameter& param,
 
 void StdAdaGradSGDRule::UpdateValueWork(float* w, float* sgd, const float* grad,
                                         float scale) {
-  for (int i = 0; i < _embedding_dim; i++) {
+  for (size_t i = 0; i < _embedding_dim; i++) {
     float& g2sum = sgd[G2SumIndex() + i];
     double scaled_grad = grad[i] / scale;
     w[i] -= learning_rate_ * scaled_grad *
@@ -151,7 +151,7 @@ void StdAdaGradSGDRule::UpdateValueWork(float* w, float* sgd, const float* grad,
 
 void StdAdaGradSGDRule::InitValueWork(float* value, float* sgd,
                                       bool zero_init) {
-  for (int i = 0; i < _embedding_dim; ++i) {
+  for (size_t i = 0; i < _embedding_dim; ++i) {
     if (zero_init) {
       value[i] = 0.0;
       BoundValue(value[i]);
@@ -202,7 +202,7 @@ void SparseAdamSGDRule::UpdateValueWork(float* w, float* sgd, const float* grad,
 
   // lr not change in one update
   lr *= sqrt(1 - beta2_pow_) / (1 - beta1_pow_);
-  for (int i = 0; i < _embedding_dim; i++) {
+  for (size_t i = 0; i < _embedding_dim; i++) {
     // Calculation
     gsum[i] = _beta1_decay_rate * gsum[i] + (1 - _beta1_decay_rate) * g[i];
     g2sum[i] =
@@ -217,7 +217,7 @@ void SparseAdamSGDRule::UpdateValueWork(float* w, float* sgd, const float* grad,
 
 void SparseAdamSGDRule::InitValueWork(float* value, float* sgd,
                                       bool zero_init) {
-  for (int i = 0; i < _embedding_dim; ++i) {
+  for (size_t i = 0; i < _embedding_dim; ++i) {
     if (zero_init) {
       value[i] = 0.0;
       BoundValue(value[i]);
@@ -231,7 +231,7 @@ void SparseAdamSGDRule::InitValueWork(float* value, float* sgd,
     }
   }
   // init rule gsum and g2sum
-  for (int i = GSumIndex(); i < Beta1PowIndex(); i++) {
+  for (size_t i = GSumIndex(); i < Beta1PowIndex(); i++) {
     sgd[i] = 0.0;
   }
   // init beta1_pow and beta2_pow
